backend/requirements.txt
# Core Dependencies
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
requests==2.31.0
Werkzeug==3.0.1
python-dotenv==1.0.1
redis==5.0.1
SQLAlchemy==2.0.28
email-validator==2.1.0.post1

# Database
psycopg2-binary==2.9.9  # For PostgreSQL support
alembic==1.13.1  # For database migrations

# Security
bcrypt==4.1.2  # For password hashing
cryptography==42.0.2  # For cryptographic operations

# Development Tools
pytest==8.0.0  # For testing
pytest-cov==4.1.0  # For test coverage
black==24.1.1  # For code formatting
flake8==7.0.0  # For code linting
mypy==1.8.0  # For type checking

# Production
gunicorn==21.2.0  # Production WSGI server
python-json-logger==2.0.7  # For structured logging

# PDF Generation
reportlab==4.1.0

# Stripe
stripe>=8.0.0
